% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brandr-package.R
\docType{package}
\name{brandr-package}
\alias{brandr}
\alias{brandr-package}
\title{brandr: Brand Identity Management Using brand.yml Standard}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A system to facilitate brand identity management using the 'brand.yml' standard, providing functions to consistently access and apply brand colors, typography, and other visual elements across your R projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/brandr/}
  \item \url{https://github.com/danielvartan/brandr/}
  \item Report bugs at \url{https://github.com/danielvartan/brandr/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
