% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand.R
\name{get_brand_color_tint}
\alias{get_brand_color_tint}
\title{Get tints of brand colors}
\usage{
get_brand_color_tint(
  position = 500,
  color = "primary",
  brand_yml = here::here("_brand.yml")
)
}
\arguments{
\item{position}{(Optional) A \code{\link[base:numeric]{numeric}} vector indicating
the position of the brand color in the range of tints. The range of
positions is from \code{0} to \code{1000} (Default: \code{500}).}

\item{color}{A \code{\link[base:character]{character}} string indicating the name
of a color present in the \code{color} section of the \verb{_brand.yml}
file (Default: \code{"primary"}).}

\item{brand_yml}{(Optional) A \code{\link[base:character]{character}} string
indicating the path to the \verb{_brand.yml} file. If not provided, the
function will look for a \verb{_brand.yml} file in the current working
directory.}
}
\value{
A \code{\link[base:character]{character}} vector with hexadecimal color
codes.
}
\description{
\code{get_brand_color_tint()} generates a range of tints (color variations) for a
specific brand color, from black (position 0) through the brand color
(position 500) to white (position 1000), and returns the hexadecimal color
code at the specified position.
}
\examples{
if (interactive()) {
  seq(0, 1000, 250)
  #> [1] 0  250  500  750 1000 # Expected

  get_brand_color_tint(seq(0, 1000, 250), color = "primary")
  #> [1] "#000000" "#6D271E" "#DA4E3C" "#ECA69D" "#FFFFFF" # Expected
}
}
\seealso{
Other utility functions.: 
\code{\link{get_brand_color}()},
\code{\link{get_brand_color_mix}()},
\code{\link{get_brand_font}()}
}
\concept{utility functions.}
