% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_brand.R
\name{color_brand_sequential}
\alias{color_brand_sequential}
\alias{color_brand_diverging}
\alias{color_brand_qualitative}
\title{Brand color functions}
\usage{
color_brand_sequential(n, direction = 1)

color_brand_diverging(n, direction = 1)

color_brand_qualitative(n, direction = 1)
}
\arguments{
\item{n}{This parameter accepts two types of inputs:
\itemize{
\item If the value is an integer number and \code{type} is \code{seq} or \code{div}, the
function will return a discrete color spectrum with \code{n} colors.
\item if the value is an integer number and \code{type} is \code{qual}, the function
will return \code{n} colors from the \code{colors} parameter, in the same order,
repeating them, if necessary.
\item If the value is a \code{\link[base:numeric]{numeric}} vector between \code{0} and \code{1},
the function will return the color positions at \code{n} considering a
continuous color spectrum ranging from \code{0} to \code{1}.
}}

\item{direction}{(Optional) Sets the order of colors in the scale. If \code{1}
colors are as output by the palette function. If \code{-1}, the order of
colors is reversed (Default: \code{1}).}
}
\value{
A \code{\link[base:character]{character}} vector with hexadecimal color
codes.
}
\description{
\verb{color_brand_*} functions are wrappers of
\code{\link[=interpolate_colors]{interpolate_colors()}} for sequential, diverging,
and qualitative brand color scales. They serve as facilitators for the
visualization of the colors being used in each brand scale.
}
\details{
To control the colors for each brand color scale, assign the desired
hexadecimal color codes in a \code{\link[base:character]{character}} vector to the
following options:
\itemize{
\item \code{BRANDR_COLOR_SEQUENTIAL} for sequential color scales
\item \code{BRANDR_COLOR_DIVERGING} for diverging color scales
\item \code{BRANDR_COLOR_QUALITATIVE} for qualitative color scales
}

Use \code{\link[=get_brand_color]{get_brand_color()}} to get the hexadecimal
color codes from the \verb{_brand.yml} file.

Example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(
  BRANDR_COLOR_SEQUENTIAL =
    get_brand_color(c("primary", "secondary")),
  BRANDR_COLOR_DIVERGING =
    get_brand_color(c("primary", "white", "secondary")),
  BRANDR_COLOR_QUALITATIVE =
    get_brand_color(c("primary", "secondary", "tertiary"))
)
}\if{html}{\out{</div>}}
}
\examples{
if (interactive()) {
  color_brand_sequential(5)
  color_brand_diverging(5)
  color_brand_qualitative(5)
}
}
\seealso{
Other color functions.: 
\code{\link{interpolate_colors}()}
}
\concept{color functions.}
