% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand.R
\name{get_brand_color}
\alias{get_brand_color}
\title{Get brand colors}
\usage{
get_brand_color(color, brand_yml = here::here("_brand.yml"))
}
\arguments{
\item{color}{A \code{\link[base:character]{character}} vector indicating the name of
colors present in the \code{color} section of the \verb{_brand.yml} file.}

\item{brand_yml}{(Optional) A \code{\link[base:character]{character}} string
indicating the path to the \verb{_brand.yml} file. If not provided, the
function will look for a \verb{_brand.yml} file in the current working
directory.}
}
\value{
A \code{\link[base:character]{character}} vector with hexadecimal color
codes.
}
\description{
\code{get_brand_color()} retrieves hexadecimal color codes from the \verb{_brand.yml}
file.
}
\examples{
if (interactive()) {
  get_brand_color("primary")
  get_brand_color("secondary")
  get_brand_color("tertiary")

  get_brand_color(c("primary", "secondary"))
  get_brand_color(c("red", "green", "blue"))
}
}
\seealso{
Other utility functions.: 
\code{\link{get_brand_color_mix}()},
\code{\link{get_brand_color_tint}()},
\code{\link{get_brand_font}()}
}
\concept{utility functions.}
