% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand.R
\name{get_brand_color_mix}
\alias{get_brand_color_mix}
\title{Get a mix of brand colors}
\usage{
get_brand_color_mix(
  position = 500,
  color_1 = "primary",
  color_2 = "secondary",
  alpha = 0.5,
  brand_yml = here::here("_brand.yml")
)
}
\arguments{
\item{position}{(Optional) A \code{\link[base:numeric]{numeric}} vector indicating
the position of the brand color in the range of tints. The range of
positions is from \code{0} to \code{1000} (Default: \code{500}).}

\item{color_1, color_2}{A \code{\link[base:character]{character}} string indicating
the name of a color present in the \code{color} section of the \verb{_brand.yml}
file.}

\item{alpha}{A number between 0 and 1 indicating the alpha transparency of
the color mix (Default: \code{0.5}).}

\item{brand_yml}{(Optional) A \code{\link[base:character]{character}} string
indicating the path to the \verb{_brand.yml} file. If not provided, the
function will look for a \verb{_brand.yml} file in the current working
directory.}
}
\value{
A \code{\link[base:character]{character}} vector with hexadecimal color
codes.
}
\description{
\code{get_brand_color_mix()} mixes two specific brand colors.
}
\examples{
if (interactive()) {
  get_brand_color_mix(
    position = 500,
    color_1 = "primary",
    color_2 = "secondary",
    alpha = 0.5
  )
}
}
\seealso{
Other utility functions.: 
\code{\link{get_brand_color}()},
\code{\link{get_brand_color_tint}()},
\code{\link{get_brand_font}()}
}
\concept{utility functions.}
