% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand.R
\name{get_brand_font}
\alias{get_brand_font}
\title{Get brand fonts/typefaces}
\usage{
get_brand_font(font, brand_yml = here::here("_brand.yml"))
}
\arguments{
\item{font}{A \code{\link[base:character]{character}} vector indicating the name of
fonts/typefaces categories present in the \code{typography} section of the
\verb{_brand.yml} file.}

\item{brand_yml}{(Optional) A \code{\link[base:character]{character}} string
indicating the path to the \verb{_brand.yml} file. If not provided, the
function will look for a \verb{_brand.yml} file in the current working
directory.}
}
\value{
A \code{\link[base:character]{character}} vector with fonts/typeface names.
}
\description{
\code{get_brand_font()} retrieves the names of fonts/typefaces in the
\verb{_brand.yml} file.
}
\examples{
if (interactive()) {
  get_brand_font("base")
  get_brand_font("headings")
  get_brand_font("monospace")
  get_brand_font("monospace-inline")
  get_brand_font("monospace-block")

  get_brand_font(c("base", "headings"))
}
}
\seealso{
Other utility functions.: 
\code{\link{get_brand_color}()},
\code{\link{get_brand_color_mix}()},
\code{\link{get_brand_color_tint}()}
}
\concept{utility functions.}
